added held item rule
