+++ include/battle_rules.h
@@ -7,51 +7,6 @@ u8 GetRandomBattleRuleSeeded(void);
 u8 GetRandomTypeSeeded(void);
 bool8 IsTruantBattleRule(u32 battler);
 
-#define B_BATTLERULE_NOSWITCHING_NOPARTYMENU    TRUE    // If TRUE, the party menu won't be shown when the switch rule activates
-
-enum battleRules {
-    BATTLERULE_NONE,
-    BATTLERULE_BANNEDTYPE,
-    BATTLERULE_BANNEDMOVETYPE,
-    BATTLERULE_NOHEALING,
-    BATTLERULE_NOCRITS,
-    BATTLERULE_NORECOIL,
-    BATTLERULE_NOSTAB,
-    BATTLERULE_ONLYSTAB,
-    BATTLERULE_NOSUPEREFFECTIVE,
-    BATTLERULE_1HP,
-    BATTLERULE_1PP,
-    BATTLERULE_NOSAMESEX,
-    BATTLERULE_NOSWITCHING,
-    BATTLERULE_PERISHCOUNT,
-    BATTLERULE_SWITCHMOVES,
-    BATTLERULE_NOPRIO,
-    BATTLERULE_NOHELDITEMS,
-    BATTLERULE_NOABILITY,
-    BATTLERULE_NOSTATUS,
-    BATTLERULE_SHAREDDAMAGE,
-    BATTLERULE_NOSETUP,
-    BATTLERULE_TRICKROOM,
-    BATTLERULE_BANNEDMOVECAT_PHYSICAL,
-    BATTLERULE_BANNEDMOVECAT_SPECIAL,
-    BATTLERULE_BANNEDMOVECAT_STATUS,
-    BATTLERULE_INVERSE,
-    BATTLERULE_FIRSTMOVEONLY,
-    BATTLERULE_TRUANT,
-    BATTLERULE_NOMISSES
-};
-
-enum battleRuleCategories {
-    BATTLERULE_CATEGORY_SENDOUT,
-    BATTLERULE_CATEGORY_USEMOVE,
-    BATTLERULE_CATEGORY_MOVEEFFECT,
-    BATTLERULE_CATEGORY_SWITCHOUT,
-    BATTLERULE_CATEGORY_ACTIVATE_ITEM,
-    BATTLERULE_CATEGORY_ACTIVATE_ABILITY,
-    BATTLERULE_CATEGORY_MULTIPLE,
-    BATTLERULE_CATEGORY_GENERAL
-};
-
 struct BattleRule
 {
     u8 weight; // currently not used;
