+++ include/constants/battle_rules.h
@@ -1,0 +1,49 @@
+#ifndef GUARD_CONSTANTS_BATTLE_RULES_H
+#define GUARD_CONSTANTS_BATTLE_RULES_H
+
+#define B_BATTLERULE_NOSWITCHING_NOPARTYMENU    TRUE    // If TRUE, the party menu won't be shown when the switch rule activates
+
+enum battleRules {
+    BATTLERULE_NONE,
+    BATTLERULE_BANNEDTYPE,
+    BATTLERULE_BANNEDMOVETYPE,
+    BATTLERULE_NOHEALING,
+    BATTLERULE_NOCRITS,
+    BATTLERULE_NORECOIL,
+    BATTLERULE_NOSTAB,
+    BATTLERULE_ONLYSTAB,
+    BATTLERULE_NOSUPEREFFECTIVE,
+    BATTLERULE_1HP,
+    BATTLERULE_1PP,
+    BATTLERULE_NOSAMESEX,
+    BATTLERULE_NOSWITCHING,
+    BATTLERULE_PERISHCOUNT,
+    BATTLERULE_SWITCHMOVES,
+    BATTLERULE_NOPRIO,
+    BATTLERULE_NOHELDITEMS,
+    BATTLERULE_NOABILITY,
+    BATTLERULE_NOSTATUS,
+    BATTLERULE_SHAREDDAMAGE,
+    BATTLERULE_NOSETUP,
+    BATTLERULE_TRICKROOM,
+    BATTLERULE_BANNEDMOVECAT_PHYSICAL,
+    BATTLERULE_BANNEDMOVECAT_SPECIAL,
+    BATTLERULE_BANNEDMOVECAT_STATUS,
+    BATTLERULE_INVERSE,
+    BATTLERULE_FIRSTMOVEONLY,
+    BATTLERULE_TRUANT,
+    BATTLERULE_NOMISSES
+};
+
+enum battleRuleCategories {
+    BATTLERULE_CATEGORY_SENDOUT,
+    BATTLERULE_CATEGORY_USEMOVE,
+    BATTLERULE_CATEGORY_MOVEEFFECT,
+    BATTLERULE_CATEGORY_SWITCHOUT,
+    BATTLERULE_CATEGORY_ACTIVATE_ITEM,
+    BATTLERULE_CATEGORY_ACTIVATE_ABILITY,
+    BATTLERULE_CATEGORY_MULTIPLE,
+    BATTLERULE_CATEGORY_GENERAL
+};
+
+#endif // GUARD_CONSTANTS_BATTLE_RULES_H
